StrongSORT_tracking/requirements.txt
# pip install -r requirements.txt

# conda install -c conda-forge jupyterlab

# base
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
#requests>=2.23.0
torch>=1.7.0
torchvision>=0.8.1
# deep_sort 
easydict
# torchreid
Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
ipdb
motmetrics
scipy
ultralytics